add e2e test for profile dump. (#230)
