+++ cmd/mesh/manifest_shared_test.go
@@ -1,0 +1,62 @@
+// Copyright 2019 Istio Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package mesh
+
+import (
+	"bytes"
+	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+)
+
+var (
+	repoRootDir string
+	testDataDir string
+)
+
+func init() {
+	wd, err := os.Getwd()
+	if err != nil {
+		panic(err)
+	}
+	repoRootDir = filepath.Join(wd, "../..")
+
+	if err := syncCharts(); err != nil {
+		panic(err)
+	}
+}
+
+func runCommand(command string) (string, error) {
+	var out bytes.Buffer
+	rootCmd := GetRootCmd(strings.Split(command, " "))
+	rootCmd.SetOutput(&out)
+
+	if err := rootCmd.Execute(); err != nil {
+		return "", err
+	}
+	return out.String(), nil
+}
+
+func syncCharts() error {
+	cmd := exec.Command(filepath.Join(repoRootDir, "scripts/run_update_charts.sh"))
+	return cmd.Run()
+}
+
+func readFile(path string) (string, error) {
+	b, err := ioutil.ReadFile(path)
+	return string(b), err
+}
