+++ cmd/mesh/testdata/profile-dump/input/demo-auth_certmanager_on.yaml
@@ -1,0 +1,10 @@
+apiVersion: install.istio.io/v1alpha2
+kind: IstioControlPlane
+spec:
+  profile: demo-auth
+  security:
+    components:
+      citadel:
+        enabled: false
+      certManager:
+        enabled: true
