+++ cmd/mesh/testdata/profile-dump/input/sds_policy_off.yaml
@@ -1,0 +1,6 @@
+apiVersion: install.istio.io/v1alpha2
+kind: IstioControlPlane
+spec:
+  profile: sds
+  policy:
+    enabled: false
